Arduino Code
This code controls the movement of the camera to scan the picture. The two motors are each connected to an EasyDriver A3967 IC, and the drivers were connected to a 12V power supply and to the Arduino. Besides the drivers, the Arduino was also connected to two limit switches that will define the frame of the picture. Our code used as base the following tutorial  https://www.brainy-bits.com/stepper-motor-with-joystick-and-limit-switches/.