Image Processing Code
For the Image algorithm, k-means clustering has been used. This technique selects “centroids”, from which we will extract clusters. The program then runs through all of the points in the image and computes the shortest distance to the next cluster. Finally, this has to be repeated until the centers are mostly stable, which signifies that the final clusters have been found.
As we are not experts at programming, for this short project, we have used the code from charlesleifer.com.