Music Algorithm
The Music algorithm primarily uses pyaudio in conjunction with the karplus-strong algorithm. We also wanted it to play 3 simultaneous sounds so we ended up using Multithreading. The program centers around the class called “Sound”, which we input a frequency, a stretch factor, and a volume. The program will then follow Karplus-strong paper on sound production:
We first determine the sample size, as we know the frequency and the sampling frequency which has been arbitrarily determined to be 8000. Next, we will create an array of random numbers from -1 to 1 for our unedited sound signal. Now we will employ the strong_decay sound algorithm from Karplus-strong. The normal Karplus-Strong Algorithm simply averages the value before with the current value for all the values in the array. This, however, will create short, and weak sounds. Thus, the strong decay algorithm uses a stretch factor, so that the algorithm only averages if the binomial probability allows it to do so. 
Finally, in conjunction with the rest of the program, we will create 3 such Class-elements, with our respective frequency, stretch factor (which will be proportional to the saturation) and volume (which is proportional to the value), and thread them together using Threading.