Creates a combined Sound
This code combines the individual sounds produced by each picture into a song. It does this by firstly cutting the first audio file into two pieces. The second piece will be merged with a piece of the next audio file with the same time length. 

This overlaps between sounds were made in order to simulate a melody, where sounds are played while the previous one is still resonating. Then these pieces are stitched together, and this process is repeated for how many individual pictures were taken. Here the pydub library was used.