Alternative Sound using Saw Shape waves
This program creates a tone, by inputting a single frequency, stretch factor and volume.

It does this by first creating a single sine wave with the desired frequency. To create a saw shape however, it also creates 2 more sine waves, with integer multiples of the desired frequency and reduced amplitdue and adding this to the initial function. This has a very physical interpretation: When dealing with instruments in fact, one does not simply play the base frequency, as in general this sounds very metallic and not pleasent. Actually when playing an instrument a lot of 'harmonic modes' are solutions to the initial conditions, that, when added together will create a good sound. 

Here we try to replicate this, by only adding the first three. The reason why we don't add more, is that when only adding sine waves, the produced sound will be very metallic, which we don't necessarily want, so we reduce it to an amount where the number of modes is still significant, but not such that it will sound too metallic. Experimentally we have determined this number to be 3.

We were not able to implement multiple sounds using this technique, as this would require us to deal with destructive interference.