Main Script to run with Arduino
The main script is the code you should run in order to perform the whole process. The code communicates with the Arduino for it to move the camera only after a picture has been taken. It takes the picture and analyzes it using the image algorithm. As the image algorithm creates an RGB color tuple, we first need to convert it to a friendlier HSV color tuple. Also, in order to normalize the value and create a unique sound we have come up with several ways to create sound in accordance with the image:

Exponential Algorithm: This is purely Music-Mathematical. That is, it uses the fundamental musical theory that pure musical notes are exponential in nature, as in that the distance between two semitones is 2^(1/12). Thus, this algorithm first normalizes the Hue value of our color and scales it, so that it spreads to 3 octaves (36 semitones). The only problem with this method is that the fundamental periodicity of the HSV color model, in that 0 and 360 hue are virtually the same. In this algorithm 0 and 360 Hue values are the same notes separated by 3 octaves, which, while not dramatic is something to keep in mind. The saturation has simply been multiplied by one and added to 1, so that the result is definitely more than one, as we will need to take the reciprocal of the stretch factor and it being less than 0 has unpleasant consequences. Finally, the volume has just been normalized and added to 0.2. Less than 20% volume, the sound is barely hearable. 

Exponential Algorithm with Thresholds: This algorithm incorportates thresholds for saturation and value, so that the frequency is dependent on all 3 parameters rather than just a single one. The idea is, that if saturation or value are on extreme hands, we simply switch octaves, to accomodate to the darkening or lightness of the image.

Linear Fitting: Finally, we also incorporated a Linear Fit, in order to fit the values for Hue with the desired values for Frequency. While less intuitive, this makes more inuitive sense for the listener.

 In the end, it tells the Arduino to stop and produces the final song. Please, here change the directory in the code for one of your choice.